consolidate quick-search into one service + self-contained hotkey to avoid esbuild JIT threshold
